requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.3
click==8.1.7
python-dotenv==1.0.1
cryptography==41.0.7
gunicorn==21.2.0 ; sys_platform != 'win32'
